% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi-functions.R
\name{calculateFisherInformation}
\alias{calculateFisherInformation}
\title{Calculate Fisher Information Matrix in the two-state model}
\usage{
calculateFisherInformation(gamma, lambda, delta, i0 = 0.316)
}
\arguments{
\item{gamma}{Rate of recovery}

\item{lambda}{Force of infection}

\item{delta}{Vector of times between observations}

\item{i0}{Initial probability of being infected (prevalence)}
}
\value{
The determinant of the Fisher Information
}
\description{
Calculates the determinant of the Fisher Information matrix in a two-state model
for a given set of parameters and time between observations.
}
