% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmi-functions.R
\name{convertSimulationsToPanelData}
\alias{convertSimulationsToPanelData}
\title{Convert model simulations to panel data}
\usage{
convertSimulationsToPanelData(simulations, sampling_times)
}
\arguments{
\item{simulations}{A list where each element contains the time series information for a single individual, and also the ID of that individual.}

\item{sampling_times}{A list where each element contain the times to sample that specific individual.}
}
\value{
A data frame containing the panel data for the entire population.
}
\description{
Convert model simulations to panel data
}
\examples{
simulations <- simulateLinearedSIS(lambda=1/60, gamma=1/20, num_individuals=100)
sampling_times <- lapply(1:num_individuals, function(i) { rep(x=1, times=365)})
panel_data <- convertSimulationsToPanelData(simulations, sampling_times)
}
