% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmi-functions.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit model using stan}
\usage{
fit_stan(data, iterations = 10000, model = tmi_model, ...)
}
\arguments{
\item{data}{Panel data to fit for, of class "tmidata" (such as that generated by the setupData function)}

\item{iterations}{Number of iterations to run each chain for}

\item{model}{Model code for stan. The default model is that for the two-state Markov chain.}

\item{...}{Other parameters which are passed to `rstan::sampling`.}
}
\value{
The stan fit.
}
\description{
Fit model using stan
}
