% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmi-functions.R
\name{generateSamplingTimesFromPanelData}
\alias{generateSamplingTimesFromPanelData}
\title{Generates sampling times to be similar to that from already collected panel data}
\usage{
generateSamplingTimesFromPanelData(panel_data, time_variable, id = id,
  num_individuals, max_time = 365)
}
\arguments{
\item{panel_data}{Panel data from which to draw the observation distribution from}

\item{time_variable}{The variable corresponding to time (unquoted)}

\item{id}{The ID variable for each individual}

\item{num_individuals}{Number of individuals to generate sampling times for}

\item{max_time}{Maximum time to observe each individual}
}
\value{
Returns a list of sampling times for each individual.
}
\description{
Generates sampling times to be similar to that from already collected panel data
}
