% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmi-functions.R
\name{getSoreDurations}
\alias{getSoreDurations}
\title{Get the per-individual sore durations}
\usage{
getSoreDurations(panel_data, formula, id = ID)
}
\arguments{
\item{panel_data}{Panel data to fit for}

\item{formula}{A formula of the form state~time, where state is the state variable (1=infectious) and time is the observation time}

\item{id}{The ID variable which identifies each individual (provide unquoted)}
}
\value{
A vector of the empricial durations of each skin sore infection, assuming continuous observation.
}
\description{
Get the per-individual sore durations
}
