% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmi-functions.R
\name{optimiseRobustExperimentalDesign}
\alias{optimiseRobustExperimentalDesign}
\title{Optimise Robust Experimental Design}
\usage{
optimiseRobustExperimentalDesign(stan_fit, delta0 = rep(10, 12),
  num_obs = 12, num_samples = 10000)
}
\arguments{
\item{stan_fit}{An object of class tmifit}

\item{delta0}{Vector containing initial guess for time between each visit}

\item{num_samples}{Number of samples to approximate the robust design integral with. Higher values
will lead the computation to take longer.}
}
\description{
Optimises the Fisher Information matrix under the ED-optimality criterion
approximating the integral using monte-carlo sampling
}
