% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmi-functions.R
\name{plot_fits}
\alias{plot_fits}
\title{Plot marginal posteriors from fitted data.}
\usage{
plot_fits(stan_fit, lambda_ = -1, gamma_ = -1)
}
\arguments{
\item{stan_fit}{The fitted result from stan}

\item{lambda}{True value of lambda (-1 if true value is unknown)}

\item{gamma}{True value of gamma (-1 if true value is unknown)}
}
\description{
Plots the marginal posteriors for both lambda and gamma, as well as 95% CIs (solid lines) and 99% CIs (dashed lines)
}
