% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmi-functions.R
\name{simulateLinearedSIS}
\alias{simulateLinearedSIS}
\title{Simulate population according to linearised SIS model}
\usage{
simulateLinearedSIS(lambda, gamma, max_time = 365, num_individuals = 1)
}
\arguments{
\item{lambda}{The force of infection}

\item{gamma}{The rate of recovery (inverse of infectious period)}

\item{max_time}{Maximum length of time (days) for which to simulate individuals for}

\item{num_individuals}{Number of individuals to seperately simulate}
}
\value{
List where each element contains the time series information of an individual.
}
\description{
Simulate population according to linearised SIS model
}
\examples{
simulations <- simulateLinearisedSIS(lambda=1/60, gamma=1/20, num_individuals=400)
}
