% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmi-functions.R
\name{setupData}
\alias{setupData}
\title{Set up panel data for fitting}
\usage{
setupData(panel_data)
}
\arguments{
\item{panel_data}{The input panel data}
}
\value{
A list containing the data for use for fitting, of class "tmidata".
}
\description{
Set up panel data for fitting
}
